# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "salus documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/doctors": {
            "get": {
                "tags": [
                    "doctors"
                ],
                "summary": "Get all doctors",
                "operationId": "947a21b8ae7d72b24b902600c8e443c9",
                "responses": {
                    "200": {
                        "description": "List of doctors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "doctors not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/doctors/{id}": {
            "get": {
                "tags": [
                    "doctors"
                ],
                "summary": "Get all doctors",
                "operationId": "dd020a547256bc4a1307488115755404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "fetch doctor"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "doctors not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/doctors/search": {
            "get": {
                "tags": [
                    "doctors"
                ],
                "summary": "search for doctors",
                "operationId": "5863ead716cd4eadef7515b492299639",
                "parameters": [
                    {
                        "name": "specialty",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "General Medicine"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "rabat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of doctors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "doctors not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/symptoms": {
            "get": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Get all symptoms for the authenticated user",
                "description": "Display a listing of the resource.",
                "operationId": "2e85564776df45ed79b7e3b8d668674a",
                "responses": {
                    "200": {
                        "description": "List of user symptoms"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Symptoms not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Create a new symptom",
                "description": "Store a newly created resource in storage.",
                "operationId": "b4bb118e280800fed14f305fd87b1e33",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "severity",
                                    "date_recorded"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "itches"
                                    },
                                    "severity": {
                                        "type": "string",
                                        "example": "mild",
                                        "enum": [
                                            "mild",
                                            "moderate",
                                            "severe"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "....."
                                    },
                                    "date_recorded": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-03-12"
                                    },
                                    "notes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "I carried it for a week, it gets stronger"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Symptom created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/symptoms/{symptom}": {
            "get": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Get one symptoms for the authenticated user",
                "description": "Display the specified resource.",
                "operationId": "6201d2d3b951ed650aecbe181220ef09",
                "parameters": [
                    {
                        "name": "symptom",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user symptoms"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Symptoms not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Update a symptom",
                "description": "Update the specified resource in storage.",
                "operationId": "53a96dd028297f2a2354086cbdabff9f",
                "parameters": [
                    {
                        "name": "symptom",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "itches"
                                    },
                                    "severity": {
                                        "type": "string",
                                        "example": "mild",
                                        "enum": [
                                            "mild",
                                            "moderate",
                                            "severe"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "....."
                                    },
                                    "date_recorded": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-03-12"
                                    },
                                    "notes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "I carried it for a week, it gets stronger"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Symptom Updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Get one symptoms for the authenticated user",
                "description": "Remove the specified resource from storage.",
                "operationId": "ab0fd4f956bb2638c42e62be3ec5e94f",
                "parameters": [
                    {
                        "name": "symptom",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user symptoms"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Symptoms not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sunctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}